refactor: update CX sell step
